Crate solana_perf

Modules

cuda_runtime

data_budget

deduper

Utility to deduplicate baches of incoming network packets.

discard

packet

The packet module defines data structures and methods to pull data from the network.

perf_libs

recycler

recycler_cache

sigverify

The sigverify module provides digital signature verification functions. By default, signatures are verified in parallel using all available CPU cores. When perf-libs are available signature verification is offloaded to the GPU.

thread

Functions

report_target_features