Struct solana_perf::cuda_runtime::PinnedVec

pub struct PinnedVec<T: Default + Clone + Sized> {
    x: Vec<T>,
    pinned: bool,
    pinnable: bool,
    recycler: Weak<RecyclerX<PinnedVec<T>>>,
}

Fields
x: Vec<T>pinned: boolpinnable: boolrecycler: Weak<RecyclerX<PinnedVec<T>>>
Implementations

impl<T: Clone + Default + Sized> PinnedVec<T>

pub fn iter(&self) -> Iter<'_, T>

pub fn iter_mut(&mut self) -> IterMut<'_, T>

pub fn capacity(&self) -> usize

impl<T: Clone + Default + Sized> PinnedVec<T>

pub fn reserve(&mut self, size: usize)

pub fn reserve_and_pin(&mut self, size: usize)

pub fn set_pinnable(&mut self)

pub fn copy_from_slice(&mut self, data: &[T])where T: Copy,

pub fn from_vec(source: Vec<T>) -> Self

pub fn with_capacity(capacity: usize) -> Self

pub fn is_empty(&self) -> bool

pub fn len(&self) -> usize

pub fn as_ptr(&self) -> *const T

pub fn as_mut_ptr(&mut self) -> *mut T

fn prepare_realloc(&mut self, new_size: usize) -> (*mut T, usize)

pub fn push(&mut self, x: T)

pub fn truncate(&mut self, size: usize)

pub fn resize(&mut self, size: usize, elem: T)

pub fn append(&mut self, other: &mut Vec<T>)

pub fn append_pinned(&mut self, other: &mut Self)

pub unsafe fn set_len(&mut self, size: usize)

Forces the length of the vector to new_len.

This is a low-level operation that maintains none of the normal invariants of the type. Normally changing the length of a vector is done using one of the safe operations instead, such as truncate, resize, extend, or clear.

Safety

new_len must be less than or equal to capacity().

The elements at old_len..new_len must be initialized.

pub fn shuffle<R: Rng>(&mut self, rng: &mut R)

fn check_ptr( &mut self, _old_ptr: *mut T, _old_capacity: usize, _from: &'static str )