Struct solana_perf::packet::PacketBatch

pub struct PacketBatch {
    packets: PinnedVec<Packet>,
}

Fields
packets: PinnedVec<Packet>
Implementations

impl PacketBatch

pub fn new(packets: Vec<Packet>) -> Self

pub fn with_capacity(capacity: usize) -> Self

pub fn new_pinned_with_capacity(capacity: usize) -> Self

pub fn new_unpinned_with_recycler( recycler: PacketBatchRecycler, capacity: usize, name: &'static str ) -> Self

pub fn new_with_recycler( recycler: PacketBatchRecycler, capacity: usize, name: &'static str ) -> Self

pub fn new_with_recycler_data( recycler: &PacketBatchRecycler, name: &'static str, packets: Vec<Packet> ) -> Self

pub fn new_unpinned_with_recycler_data_and_dests<T: Serialize>( recycler: PacketBatchRecycler, name: &'static str, dests_and_data: &[(SocketAddr, T)] ) -> Self

pub fn new_unpinned_with_recycler_data( recycler: &PacketBatchRecycler, name: &'static str, packets: Vec<Packet> ) -> Self

pub fn resize(&mut self, new_len: usize, value: Packet)

pub fn truncate(&mut self, len: usize)

pub fn push(&mut self, packet: Packet)

pub fn set_addr(&mut self, addr: &SocketAddr)

pub fn len(&self) -> usize

pub fn capacity(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn as_ptr(&self) -> *const Packet

pub fn iter(&self) -> Iter<'_, Packet>

pub fn iter_mut(&mut self) -> IterMut<'_, Packet>

pub unsafe fn set_len(&mut self, new_len: usize)

See Vector::set_len() for more details

Safety

new_len must be less than or equal to [self.capacity].

The elements at old_len..new_len must be initialized. Packet data will likely be overwritten when populating the packet, but the meta should specifically be initialized to known values.