solana_perf

Module sigverify

The sigverify module provides digital signature verification functions. By default, signatures are verified in parallel using all available CPU cores. When perf-libs are available signature verification is offloaded to the GPU.

Enums

PacketError

Constants

VERIFY_PACKET_CHUNK_SIZE

Functions

check_packed_ge_small_order

copy_return_values

count_discarded_packets

count_packets_in_batches

count_valid_packets

ed25519_verify

ed25519_verify_cpu

ed25519_verify_disabled

generate_offsets

get_checked_scalar

init

mark_disabled

shrink_batches

Type Aliases

TxOffset