Module solana_perf::sigverify

The sigverify module provides digital signature verification functions. By default, signatures are verified in parallel using all available CPU cores. When perf-libs are available signature verification is offloaded to the GPU.

Structs

PAR_THREAD_POOL 🔒

PacketOffsets 🔒

Enums

PacketError

Constants

TRACER_KEY 🔒

TRACER_KEY_BYTES 🔒

TRACER_KEY_OFFSET_IN_TRANSACTION 🔒

VERIFY_MIN_PACKETS_PER_THREAD

Functions

__stability 🔒

__static_ref_initialize 🔒

check_for_simple_vote_transaction 🔒

check_for_tracer_packet

check_packed_ge_small_order

copy_return_values

count_discarded_packets

count_packets_in_batches

count_valid_packets

do_get_packet_offsets 🔒

ed25519_verify

ed25519_verify_cpu

ed25519_verify_disabled

generate_offsets

get_checked_scalar

get_packet_offsets 🔒

init

mark_disabled

shrink_batches

verify_packet 🔒

Returns true if the signatrue on the packet verifies. Caller must do packet.set_discard(true) if this returns false.

Type Definitions

TxOffset

TxOffsets 🔒